refactor: rename data into particles
