+++ expertsystem/amplitude/canonical_decay.py
@@ -2,7 +2,7 @@
 
 from typing import Any, Callable, List, Optional
 
-from expertsystem.data import Spin
+from expertsystem.particles import Spin
 from expertsystem.solving.topology import StateTransitionGraph
 from expertsystem.solving.types import InteractionProperties, ParticleWithSpin
