+++ expertsystem/io/yaml/__init__.py
@@ -8,7 +8,7 @@ __all__ = [
 import yaml
 
 from expertsystem.amplitude.model import AmplitudeModel
-from expertsystem.data import Particle, ParticleCollection, Spin
+from expertsystem.particles import Particle, ParticleCollection, Spin
 
 from . import _build, _dump
