+++ expertsystem/io/yaml/_build.py
@@ -25,7 +25,7 @@ from expertsystem.amplitude.model import (
     SequentialAmplitude,
     StrengthIntensity,
 )
-from expertsystem.data import Parity, Particle, ParticleCollection, Spin
+from expertsystem.particles import Parity, Particle, ParticleCollection, Spin
 
 from . import validation
