+++ expertsystem/particles.py
