+++ tests/unit/amplitude/test_model.py
@@ -10,7 +10,7 @@ from expertsystem.amplitude.model import (
     RelativisticBreitWigner,
     _assert_arg_type,
 )
-from expertsystem.data import ParticleCollection
+from expertsystem.particles import ParticleCollection
 
 
 class TestFitParameters:
