+++ tests/unit/io/test_particle_list.py
@@ -2,7 +2,7 @@
 import pytest
 
 from expertsystem import io
-from expertsystem.data import ParticleCollection
+from expertsystem.particles import ParticleCollection
 
 XML_FILE = "particle_list.xml"
 YAML_FILE = "particle_list.yml"
