+++ tests/unit/solving/conservation_rules/test_spin.py
@@ -2,7 +2,7 @@ from typing import List, Tuple
 
 import pytest
 
-from expertsystem.data import Spin
+from expertsystem.particles import Spin
 from expertsystem.solving.conservation_rules import (
     IsoSpinConservation,
     IsoSpinEdgeInput,
